Refactor code into the abstract DeleteStudent class
